requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0
python-multipart>=0.0.5
pydantic>=1.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
streamlit>=1.32.0
requests>=2.31.0
